backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose
python-dotenv
redis
fastapi-admin
